chore: move test from integration-tests to connect-popup
