+++ docker/docker-compose.connect-popup-test.yml
@@ -31,7 +31,7 @@ services:
       # useful for debugging tests
       - PWDEBUG=console
     working_dir: /trezor-suite
-    command: bash -c "npx playwright install && docker/wait-for-200.sh http://localhost:8088/index.html && yarn workspace @trezor/integration-tests test:connect-popup"
+    command: bash -c "npx playwright install && docker/wait-for-200.sh http://localhost:8088/index.html && yarn workspace @trezor/connect-popup test:e2e"
     volumes:
       - ../:/trezor-suite
       - /tmp/.X11-unix:/tmp/.X11-unix:rw
