+++ packages/connect-popup/e2e/tests/__fixtures__/methods.js
