+++ packages/connect-popup/package.json
@@ -7,7 +7,8 @@
         "lint:js": "yarn g:eslint '**/*.{ts,tsx,js}'",
         "type-check": "tsc --build",
         "dev": "rimraf build && TS_NODE_PROJECT=\"tsconfig.json\" yarn webpack --config ./webpack/dev.webpack.config.ts",
-        "build": "rimraf build && TS_NODE_PROJECT=\"tsconfig.json\" yarn webpack --config ./webpack/prod.webpack.config.ts"
+        "build": "rimraf build && TS_NODE_PROJECT=\"tsconfig.json\" yarn webpack --config ./webpack/prod.webpack.config.ts",
+        "test:e2e": "yarn workspace @trezor/integration-tests wait-for-env && yarn xvfb-maybe -- playwright test --config=./e2e/playwright.config.ts"
     },
     "dependencies": {
         "@trezor/components": "*",
@@ -18,13 +19,16 @@
         "styled-components": "^5.3.5"
     },
     "devDependencies": {
+        "@playwright/test": "^1.22.2",
         "@types/react": "^17.0.0",
         "@types/styled-components": "^5.1.25",
         "jest": "^26.6.3",
+        "playwright": "^1.22.2",
         "rimraf": "^3.0.2",
         "typescript": "4.7.4",
         "webpack": "^5.73.0",
         "webpack-cli": "^4.10.0",
-        "webpack-dev-server": "^4.7.4"
+        "webpack-dev-server": "^4.7.4",
+        "xvfb-maybe": "^0.2.1"
     }
 }
