--- packages/integration-tests/projects/connect-popup/env.js
@@ -1,3 +1,0 @@
-import fs from 'fs';
-import path from 'path';
-import { Controller } from '../../websocket-client';
